Update User entity to work with FOSUserBundle
