+++ app/config/security.yml
@@ -5,8 +5,9 @@ security:
         FOS\UserBundle\Model\UserInterface: bcrypt
 
     role_hierarchy:
-        ROLE_ADMIN:       ROLE_USER
-        ROLE_SUPER_ADMIN: ROLE_ADMIN
+        ROLE_USER:    ROLE_USER
+        ROLE_AUDITOR: ROLE_AUDITOR
+        ROLE_OWNER:   ROLE_OWNER
 
     providers:
         fos_userbundle:
