--- src/MMProjectBundle/Entity/Resources.php
@@ -1,67 +1,0 @@
-<?php
-
-namespace MMProjectBundle\Entity;
-
-use Doctrine\ORM\Mapping as ORM;
-
-/**
- * Resources
- *
- * @ORM\Table(name="resources")
- * @ORM\Entity
- */
-class Resources
-{
-    /**
-     * @var integer
-     *
-     * @ORM\Column(name="id", type="integer")
-     * @ORM\Id
-     * @ORM\GeneratedValue(strategy="IDENTITY")
-     */
-    private $id;
-
-    /**
-     * @var string
-     *
-     * @ORM\Column(name="name", type="string", length=255, nullable=false)
-     */
-    private $name;
-
-    /**
-     * @return int
-     */
-    public function getId(): int
-    {
-        return $this->id;
-    }
-
-    /**
-     * @param int $id
-     * @return Resources
-     */
-    public function setId(int $id): Resources
-    {
-        $this->id = $id;
-        return $this;
-    }
-
-    /**
-     * @return string
-     */
-    public function getName(): string
-    {
-        return $this->name;
-    }
-
-    /**
-     * @param string $name
-     * @return Resources
-     */
-    public function setName(string $name): Resources
-    {
-        $this->name = $name;
-        return $this;
-    }
-}
-
