--- src/MMProjectBundle/Entity/Users.php
@@ -1,387 +1,0 @@
-<?php
-
-namespace MMProjectBundle\Entity;
-
-use Doctrine\ORM\Mapping as ORM;
-use FOS\UserBundle\Entity\User as BaseUser;
-
-/**
- * Users
- *
- * @ORM\Table(name="users", uniqueConstraints={@ORM\UniqueConstraint(name="users_email_uindex", columns={"email"})})
- * @ORM\Entity
- */
-class Users extends BaseUser
-{
-    /**
-     * @var integer
-     *
-     * @ORM\Column(name="id", type="bigint")
-     * @ORM\Id
-     * @ORM\GeneratedValue(strategy="IDENTITY")
-     */
-    protected $id;
-
-    /**
-     * @var integer
-     *
-     * @ORM\Column(name="version", type="integer", nullable=false)
-     */
-    private $version;
-
-    /**
-     * @var \DateTime
-     *
-     * @ORM\Column(name="date_created", type="datetime", nullable=false)
-     */
-    private $dateCreated;
-
-    /**
-     * @var \DateTime
-     *
-     * @ORM\Column(name="last_updated", type="datetime", nullable=false)
-     */
-    private $lastUpdated;
-
-    /**
-     * @var string
-     *
-     * @ORM\Column(name="first_name", type="string", length=255, nullable=false)
-     */
-    private $firstName;
-
-    /**
-     * @var string
-     *
-     * @ORM\Column(name="last_name", type="string", length=255, nullable=false)
-     */
-    private $lastName;
-
-    /**
-     * @var string
-     *
-     * @ORM\Column(name="phone_number", type="string", length=15, nullable=false)
-     */
-    private $phoneNumber;
-
-    /**
-     * @var string
-     *
-     * @ORM\Column(name="zipcode", type="string", length=6, nullable=true)
-     */
-    private $zipcode;
-
-    /**
-     * @var string
-     *
-     * @ORM\Column(name="address", type="string", length=255, nullable=true)
-     */
-    private $address;
-
-    /**
-     * @var string
-     *
-     * @ORM\Column(name="street_with_no", type="string", length=255, nullable=true)
-     */
-    private $streetWithNo;
-
-    /**
-     * @var string
-     *
-     * @ORM\Column(name="country_code", type="string", length=4, nullable=true)
-     */
-    private $countryCode;
-
-    /**
-     * @var string
-     *
-     * @ORM\Column(name="city", type="string", length=45, nullable=true)
-     */
-    private $city;
-
-    /**
-     * @var string
-     *
-     * @ORM\Column(name="status", type="string", length=45, nullable=true)
-     */
-    private $status;
-
-    /**
-     * @var \Doctrine\Common\Collections\Collection
-     *
-     * @ORM\ManyToMany(targetEntity="MMProjectBundle\Entity\Roles", mappedBy="user")
-     */
-    private $role;
-
-    /**
-     * Constructor
-     */
-    public function __construct()
-    {
-        parent::__construct();
-        $this->role = new \Doctrine\Common\Collections\ArrayCollection();
-        $this->dateCreated = new \DateTime();
-        $this->lastUpdated = new \DateTime();
-    }
-
-    /**
-     * @ORM\PreUpdate()
-     */
-    public function onPreUpdate()
-    {
-        $this->lastUpdated = new \DateTime();
-    }
-
-    /**
-     * @return int
-     */
-    public function getId(): int
-    {
-        return $this->id;
-    }
-
-    /**
-     * @param int $id
-     * @return Users
-     */
-    public function setId(int $id): Users
-    {
-        $this->id = $id;
-        return $this;
-    }
-
-    /**
-     * @return int
-     */
-    public function getVersion(): int
-    {
-        return $this->version;
-    }
-
-    /**
-     * @param int $version
-     * @return Users
-     */
-    public function setVersion(int $version): Users
-    {
-        $this->version = $version;
-        return $this;
-    }
-
-    /**
-     * @return \DateTime
-     */
-    public function getDateCreated(): \DateTime
-    {
-        return $this->dateCreated;
-    }
-
-    /**
-     * @param \DateTime $dateCreated
-     * @return Users
-     */
-    public function setDateCreated(\DateTime $dateCreated): Users
-    {
-        $this->dateCreated = $dateCreated;
-        return $this;
-    }
-
-    /**
-     * @return \DateTime
-     */
-    public function getLastUpdated(): \DateTime
-    {
-        return $this->lastUpdated;
-    }
-
-    /**
-     * @param \DateTime $lastUpdated
-     * @return Users
-     */
-    public function setLastUpdated(\DateTime $lastUpdated): Users
-    {
-        $this->lastUpdated = $lastUpdated;
-        return $this;
-    }
-
-    /**
-     * @return string
-     */
-    public function getFirstName(): string
-    {
-        return $this->firstName;
-    }
-
-    /**
-     * @param string $firstName
-     * @return Users
-     */
-    public function setFirstName(string $firstName): Users
-    {
-        $this->firstName = $firstName;
-        return $this;
-    }
-
-    /**
-     * @return string
-     */
-    public function getLastName(): string
-    {
-        return $this->lastName;
-    }
-
-    /**
-     * @param string $lastName
-     * @return Users
-     */
-    public function setLastName(string $lastName): Users
-    {
-        $this->lastName = $lastName;
-        return $this;
-    }
-
-    /**
-     * @return string
-     */
-    public function getPhoneNumber(): string
-    {
-        return $this->phoneNumber;
-    }
-
-    /**
-     * @param string $phoneNumber
-     * @return Users
-     */
-    public function setPhoneNumber(string $phoneNumber): Users
-    {
-        $this->phoneNumber = $phoneNumber;
-        return $this;
-    }
-
-    /**
-     * @return string
-     */
-    public function getZipcode(): string
-    {
-        return $this->zipcode;
-    }
-
-    /**
-     * @param string $zipcode
-     * @return Users
-     */
-    public function setZipcode(string $zipcode): Users
-    {
-        $this->zipcode = $zipcode;
-        return $this;
-    }
-
-    /**
-     * @return string
-     */
-    public function getAddress(): string
-    {
-        return $this->address;
-    }
-
-    /**
-     * @param string $address
-     * @return Users
-     */
-    public function setAddress(string $address): Users
-    {
-        $this->address = $address;
-        return $this;
-    }
-
-    /**
-     * @return string
-     */
-    public function getStreetWithNo(): string
-    {
-        return $this->streetWithNo;
-    }
-
-    /**
-     * @param string $streetWithNo
-     * @return Users
-     */
-    public function setStreetWithNo(string $streetWithNo): Users
-    {
-        $this->streetWithNo = $streetWithNo;
-        return $this;
-    }
-
-    /**
-     * @return string
-     */
-    public function getCountryCode(): string
-    {
-        return $this->countryCode;
-    }
-
-    /**
-     * @param string $countryCode
-     * @return Users
-     */
-    public function setCountryCode(string $countryCode): Users
-    {
-        $this->countryCode = $countryCode;
-        return $this;
-    }
-
-    /**
-     * @return string
-     */
-    public function getCity(): string
-    {
-        return $this->city;
-    }
-
-    /**
-     * @param string $city
-     * @return Users
-     */
-    public function setCity(string $city): Users
-    {
-        $this->city = $city;
-        return $this;
-    }
-
-    /**
-     * @return string
-     */
-    public function getStatus(): string
-    {
-        return $this->status;
-    }
-
-    /**
-     * @param string $status
-     * @return Users
-     */
-    public function setStatus(string $status): Users
-    {
-        $this->status = $status;
-        return $this;
-    }
-
-    /**
-     * @return \Doctrine\Common\Collections\Collection
-     */
-    public function getRole(): \Doctrine\Common\Collections\Collection
-    {
-        return $this->role;
-    }
-
-    /**
-     * @param \Doctrine\Common\Collections\Collection $role
-     * @return Users
-     */
-    public function setRole(\Doctrine\Common\Collections\Collection $role): Users
-    {
-        $this->role = $role;
-        return $this;
-    }
-}
-
